+++ collect/tracehound-collect.sh
@@ -34,7 +34,7 @@
 
 set -eu
 
-VERSION="0.9.2"
+VERSION="0.9.3"
 OUT=""
 REFERENCE=""
 HOSTNAME_OVERRIDE=""